+++ src/graphn/_generated/api/secrets/create_secret.py
@@ -52,6 +52,11 @@ def _parse_response(
 
         return response_401
 
+    if response.status_code == 403:
+        response_403 = Error.from_dict(response.json())
+
+        return response_403
+
     if response.status_code == 409:
         response_409 = Error.from_dict(response.json())